Version bump to 0.9.9.5
Fixed _pick so that if using column names will not select column one to
the left of the one desired.

Reworked unstack a bit to try and adress memory problems.

Fixed test_ts_intervals so that it at least would test the PANDAS codes
if not the the tstep on Python 2.6 and 3.3